Move ReportParser to sendletter package
