+++ src/main/java/uk/gov/hmcts/reform/sendletter/services/ReportParsingException.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.slc.services;
+package uk.gov.hmcts.reform.sendletter.services;
 
 public class ReportParsingException extends RuntimeException {
     public ReportParsingException(Throwable cause) {
