+++ src/test/java/uk/gov/hmcts/reform/sendletter/services/ReportParserTest.java
@@ -2,8 +2,6 @@ package uk.gov.hmcts.reform.sendletter.services;
 
 import org.junit.Test;
 import uk.gov.hmcts.reform.slc.model.LetterPrintStatus;
-import uk.gov.hmcts.reform.slc.services.ReportParser;
-import uk.gov.hmcts.reform.slc.services.ReportParsingException;
 import uk.gov.hmcts.reform.slc.services.steps.sftpupload.ParsedReport;
 import uk.gov.hmcts.reform.slc.services.steps.sftpupload.Report;
